Create forEach.js
Part 1

Create a function forEach which takes an array and a callback, and runs the callback on each element of the array. forEach does not return anything. Please do not use the native forEach or map method.

Part 2

Now let's rebuild map from the previous challenge. This time instead of using a for loop, you're going to use the forEach we just created.